Corrigindo erros de PR's anteriores

+++ sql.py
@@ -2,7 +2,6 @@ import sqlite3
 import sys
 db = 'banco.db'
 
-
 def refazer(table):
 	conn = sqlite3.connect(db)
 	cursor = conn.cursor()
@@ -10,72 +9,53 @@ def refazer(table):
 		cursor.execute('DROP TABLE {};'.format(table))
 		conn.commit()
 		return 'restaurado'
-	except BaseException:
+	except:
 		return 'erro ao refazer'
-
 
 def criar_table(table):
 	conn = sqlite3.connect(db)
 	cursor = conn.cursor()
 	try:
-		cursor.execute(
-			"""CREATE TABLE IF NOT EXISTS {}(nome VARCHAR(50) NOT NULL, advs INT NOT NULL DEFAULT 0, Id INTEGER PRIMARY KEY AUTOINCREMENT, user_id INT, alert INT DEFAULT 0);""".format(
-				str(table).replace(
-					'-',
-					'T')))
+		cursor.execute("""CREATE TABLE IF NOT EXISTS {}(nome VARCHAR(50) NOT NULL, advs INT NOT NULL DEFAULT 0, Id INTEGER PRIMARY KEY AUTOINCREMENT, user_id INT, alert INT DEFAULT 0);""".format(str(table).replace('-', 'T')))
 		conn.commit()
 		return 'Table {} criado'.format(str(table).replace('-', 'T'))
-	except BaseException:
+	except:
 		return 'erro ao criar table {}'.format(table)
 		exit()
 	conn.close()
-
 
 def alterar_table(table):
 	conn = sqlite3.connect(db)
 	cursor = conn.cursor()
 	try:
-		cursor.execute(
-			"ALTER TABLE {} RENAME TO sqlitestudio_temp_table;".format(
-				str(table)))
+		cursor.execute("ALTER TABLE {} RENAME TO sqlitestudio_temp_table;".format(str(table)))
 		conn.commit()
-		cursor.execute(
-			"CREATE TABLE {} (nome VARCHAR (50) NOT NULL, advs INT NOT NULL DEFAULT 0, Id INTEGER PRIMARY KEY AUTOINCREMENT, user_id INT, alert INT DEFAULT 0);".format(
-				str(table)))
+		cursor.execute("CREATE TABLE {} (nome VARCHAR (50) NOT NULL, advs INT NOT NULL DEFAULT 0, Id INTEGER PRIMARY KEY AUTOINCREMENT, user_id INT, alert INT DEFAULT 0);".format(str(table)))
 		conn.commit()
-		cursor.execute(
-			"INSERT INTO {} (nome, advs) SELECT nome, advs FROM sqlitestudio_temp_table;".format(
-				str(table)))
+		cursor.execute("INSERT INTO {} (nome, advs) SELECT nome, advs FROM sqlitestudio_temp_table;".format(str(table)))
 		conn.commit()
 		cursor.execute("DROP TABLE sqlitestudio_temp_table;")
 		conn.commit()
 		print('banco alterado com sucesso')
 		conn.close()
-	except BaseException:
+	except:
 		print('erro')
-
 
 def inserir(table, nome, user_id):
 	conn = sqlite3.connect(db)
 	cursor = conn.cursor()
 	try:
-		cursor.execute(
-			"INSERT INTO {} (nome, user_id) VALUES ('{}','{}');".format(
-				str(table).replace(
-					'-', 'T'), str(nome), str(user_id)))
+		cursor.execute("INSERT INTO {} (nome, user_id) VALUES ('{}','{}');".format(str(table).replace('-', 'T'), str(nome), str(user_id)))
 		conn.commit()
 		return 'inserido'
-	except BaseException:
+	except:
 		retorno = 'erro ao inserir'
 		try:
-			alterar_table(format(str(table).replace('-', 'T')))
-			cursor.execute(
-				"INSERT INTO {} (nome, user_id) VALUES ('{}','{}');".format(
-					str(table).replace(
-						'-', 'T'), str(nome), str(user_id)))
+			alterar_table(format(str(table).replace('-','T')))
+			cursor.execute("INSERT INTO {} (nome, user_id) VALUES ('{}','{}');".format(str(table).replace('-', 'T'), str(nome), str(user_id)))
 			conn.commit()
 			retorno = 'Tabela alterada. Dados inseridos'
-		except BaseException:
+		except:
 			retorno = 'não foi possível alterar tabela'
 		return retorno
 	conn.close()
@@ -91,34 +71,26 @@ def mostrar(table):
 			resposta += '\n|--{}'.format(nome)
 		resposta += '\n'
 		return resposta
-	except BaseException:
+	except:
 		return 'erro ao buscar.'
 	conn.close()
 
-
 def delete(table, user_id):
-
 	conn = sqlite3.connect(db)
 	cursor = conn.cursor()
 	try:
-		cursor.execute(
-			"DELETE FROM {} WHERE user_id = {};".format(
-				str(table).replace(
-					'-', 'T'), user_id))
+		cursor.execute("DELETE FROM {} WHERE user_id = {};".format(str(table).replace('-', 'T'), user_id))
 		conn.commit()
 		return 'deletado'
-	except BaseException:
+	except:
 		return 'erro ao deletar'
 	conn.close()
 
-
 def procurarUserNome(table, nome):
-
 	conn = sqlite3.connect(db)
 	cursor = conn.cursor()
 	try:
-		cursor.execute(
-			"SELECT * FROM {} WHERE nome = '{}';".format(str(table).replace('-', 'T'), nome))
+		cursor.execute("SELECT * FROM {} WHERE nome = '{}';".format(str(table).replace('-', 'T'), nome))
 		for busca in cursor.fetchall():
 			user = busca[0]
 			advs = busca[1]
@@ -126,18 +98,15 @@ def procurarUserNome(table, nome):
 			alerta = busca[4]
 		cadastro = [user, advs, user_id, alerta]
 		return cadastro
-	except BaseException:
+	except:
 		return('erro ao procurar')
 	conn.close()
-
 
 def procurar(table, user_id):
-
 	conn = sqlite3.connect(db)
 	cursor = conn.cursor()
 	try:
-		cursor.execute(
-			"SELECT * FROM {} WHERE user_id = '{}';".format(str(table).replace('-', 'T'), user_id))
+		cursor.execute("SELECT * FROM {} WHERE user_id = '{}';".format(str(table).replace('-', 'T'), user_id))
 		for busca in cursor.fetchall():
 			user = busca[0]
 			advs = busca[1]
@@ -145,86 +114,67 @@ def procurar(table, user_id):
 			alerta = busca[4]
 		cadastro = [user, advs, user_id, alerta]
 		return cadastro
-	except BaseException:
+	except:
 		return('erro ao procurar')
 	conn.close()
-
 
 def alerta(table, user_id):
 	conn = sqlite3.connect(db)
 	cursor = conn.cursor()
 	try:
-		cursor.execute(
-			"UPDATE {} SET alert=1 WHERE user_id={}".format(
-				str(table).replace(
-					'-', 'T'), user_id))
+		cursor.execute("UPDATE {} SET alert=1 WHERE user_id={}".format(str(table).replace('-','T'),user_id))
 		conn.commit()
 		conn.close()
-	except BaseException:
+	except:
 		return 'erro ao inserir alerta'
-
 
 def remAlerta(table, user_id):
 	conn = sqlite3.connect(db)
 	cursor = conn.cursor()
 	try:
-		cursor.execute(
-			"UPDATE {} SET alert=0 WHERE user_id={}".format(
-				str(table).replace(
-					'-', 'T'), user_id))
+		cursor.execute("UPDATE {} SET alert=0 WHERE user_id={}".format(str(table).replace('-','T'),user_id))
 		conn.commit()
 		conn.close()
-	except BaseException:
+	except:
 		return 'erro ao remover alerta'
 
-
 def advertir(table, user_id):
-
 	conn = sqlite3.connect(db)
 	cursor = conn.cursor()
 	try:
 		advs = procurar(table, user_id)[1]
 		advs += 1
-		cursor.execute(
-			"""UPDATE {} SET advs = {} WHERE user_id = {};""".format(
-				str(table).replace(
-					'-', 'T'), advs, user_id))
+		cursor.execute("""UPDATE {} SET advs = {} WHERE user_id = {};""".format(str(table).replace('-', 'T'), advs, user_id))
 		conn.commit()
 		return 'alterado'
-	except BaseException:
+	except:
 		return 'erro ao alterar'
 	conn.close
 
-
 def desadvertir(table, user_id, quantidade):
-
 	conn = sqlite3.connect(db)
 	cursor = conn.cursor()
 	try:
 		advs = procurar(table, user_id)[1]
 		advs -= quantidade
-		cursor.execute(
-			"""UPDATE {} SET advs = {} WHERE user_id = {};""".format(
-				str(table).replace(
-					'-', 'T'), advs, user_id))
+		cursor.execute("""UPDATE {} SET advs = {} WHERE user_id = {};""".format(str(table).replace('-', 'T'), advs, user_id))
 		conn.commit()
 		return 'alterado'
-	except BaseException:
+	except:
 		return 'erro ao alterar'
 	conn.close()
 
 
 def tabelas():
-
 	conn = sqlite3.connect(db)
 	cursor = conn.cursor()
 	tabelas = ''
 	try:
 		cursor.execute('SELECT name FROM sqlite_master;')
 		for tabela in cursor.fetchall():
-			tabelas += '\n--' + tabela[0]
-		tabelas += '\n'
+			tabelas+='\n--'+tabela[0]
+		tabelas+='\n'
 		return tabelas
-	except BaseException:
+	except:
 		return 'erro ao achar tabelas'
 	conn.close()
